Moved MetadataRefreshScheduler to Metadata directory. - Fixes #167
