+++ Kentor.AuthServices.Tests/Kentor.AuthServices.Tests.csproj
@@ -199,7 +199,7 @@
     <Compile Include="StubFactory.cs" />
     <Compile Include="WebSSO\AuthServicesUrlsTests.cs" />
     <Compile Include="Configuration\KentorAuthServicesSectionTests.cs" />
-    <Compile Include="MetadataRefreshSchedulerTests.cs" />
+    <Compile Include="Metadata\MetadataRefreshSchedulerTests.cs" />
     <Compile Include="Internal\XmlHelpersTests.cs" />
     <Compile Include="WebSSO\LogoutCommandTests.cs" />
     <Compile Include="WebSSO\StubSaml2Binding.cs" />
