+++ Kentor.AuthServices/Metadata/MetadataRefreshScheduler.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Kentor.AuthServices
+namespace Kentor.AuthServices.Metadata
 {
     internal static class MetadataRefreshScheduler
     {
